Bugfix/AX-7447 no rt on upgrade (#5584)
* check installer file

* mount dir as volume

* change dir name

* doc

* check for locks dir

* fix imports

* formatting

* change install lock path

+++ axonius-libs/src/libs/axonius-py/axonius/utils/host_utils.py
@@ -2,10 +2,11 @@ import shutil
 import time
 from pathlib import Path
 
-PYTHON_INSTALLER_LOCK_FILE = Path('/tmp/python_installer.lock')
-PYTHON_UPGRADE_LOCK_FILE = Path('/tmp/upgrade.lock')
-GUIALIVE_WATCHDOG_IN_PROGRESS = Path('/tmp/guialive_watchdog_in_progress.lock')
-MONGOALIVE_WATCHDOG_IN_PROGRESS = Path('/tmp/mongoalive_watchdog_in_progress.lock')
+PYTHON_LOCKS_DIR = Path('/tmp/ax-locks/')
+PYTHON_INSTALLER_LOCK_FILE = PYTHON_LOCKS_DIR / 'python_installer.lock'
+PYTHON_UPGRADE_LOCK_FILE = PYTHON_LOCKS_DIR / 'upgrade.lock'
+GUIALIVE_WATCHDOG_IN_PROGRESS = PYTHON_LOCKS_DIR / 'guialive_watchdog_in_progress.lock'
+MONGOALIVE_WATCHDOG_IN_PROGRESS = PYTHON_LOCKS_DIR / 'mongoalive_watchdog_in_progress.lock'
 LOCK_TOO_OLD_THRESH = 5 * 60 * 60
 
 
@@ -23,15 +24,26 @@ def get_free_disk_space():
     return shutil.disk_usage('/').free
 
 
-def check_installer_locks():
-    return check_lock_file(PYTHON_INSTALLER_LOCK_FILE) or check_lock_file(PYTHON_UPGRADE_LOCK_FILE)
+def check_installer_locks(unlink: bool = True):
+    return check_lock_file(PYTHON_INSTALLER_LOCK_FILE, unlink=unlink) or\
+        check_lock_file(PYTHON_UPGRADE_LOCK_FILE, unlink=unlink)
 
 
 def check_watchdog_action_in_progress():
     return check_lock_file(GUIALIVE_WATCHDOG_IN_PROGRESS) or check_lock_file(MONGOALIVE_WATCHDOG_IN_PROGRESS)
 
 
-def check_lock_file(path):
+def create_lock_file(lock_filename: Path):
+    try:
+        if not PYTHON_LOCKS_DIR.is_dir():
+            PYTHON_LOCKS_DIR.mkdir(exist_ok=True)
+        lock_filename.touch()
+        return True
+    except Exception:
+        return False
+
+
+def check_lock_file(path: Path, unlink: bool = True):
     """
     checks if lock file exists and not older than LOCK_TOO_OLD_THRESH. If it is, it deletes it.
     :return:
@@ -39,6 +51,11 @@ def check_lock_file(path):
     if not path.is_file():
         return False
     if time.time() - path.stat().st_ctime > LOCK_TOO_OLD_THRESH:
-        path.unlink()
+        if unlink:
+            try:
+                path.unlink()
+            except Exception:
+                # lock is too old but the file cannot be deleted.
+                return False
         return False
     return True
